Hook up Edit Profile to Api
Aded functionality to the edit profile form. It now properly submits
updated user and client/sitter data to the backend. And it correctly
prepopulates the form with their current data. Had to rejigger the way
user and set were being set across the component hierarchy to make this
work correctly.