Fix Bug Where User Could Log In As Client/Sitter Without Account
User could sign in as client or sitter even if they didn't have a
corresponding client or sitter account. Now attempting to do so throws
an error. I updated messages.js so the sign-in error informs them that
attempting to sign in to an account you don't have throws an error.