Begin Adding Message Center/Refactor Favorites
Created the files that will eventually house the message center users
will use to contact each other. Also simplified the way the front end
gets a list of all of the current user's favorites. Instead of making a
separate index request to the API, the favorites info is returned as
part of the JSON response when the user signs in.